Add files via upload
I changed the css from a separate file to the html file because I foundet more simple then I made the first step to make a 3*3 grid. I will now try to make it more advanced by adding lines in the css and put the items in the center and changing some colors.