Add files via upload
before this I was working on different java script file and I styled it in a css sheet. now I put it all in one file which is html file. i started to put some functions about the turn and I changed the class to id and right now I am facing a problem in table. it is not showing and I can not see anything. I will search the problem and fix it.